[PF-71] Move campaign-tutorial file into static-data folder
