+++ mon-pix/app/static-data/campaign-tutorial.js
@@ -1,25 +1,21 @@
 export default {
   tutorial: [
     {
-      id: 0,
       title: 'Vous pouvez rechercher sur internet',
       icon: 'icn-recherche.svg',
       explanation: 'Si vous ignorez une réponse, \nelle se trouve sûrement sur internet.'
     },
     {
-      id: 1,
       title: 'Pas de limite de temps !',
       icon: 'icn-temps.svg',
       explanation: 'Prenez le temps nécessaire pour terminer votre parcours. \nSi une question est chronométrée, cela vous sera indiqué.'
     },
     {
-      id: 2,
       title: 'Des tutos pour apprendre',
       icon: 'icn-tutos.svg',
       explanation: 'Accédez à des tutos pour apprendre davantage \nsur chaque question et progresser.'
     },
     {
-      id: 3,
       title: 'Un niveau de difficulté adapté',
       icon: 'icn-algo.svg',
       explanation: 'En fonction de vos réponses,\nPix adapte la difficulté des questions. '
